+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 ### Improvements
 - Update checks now support beta versions of the modpack
+- Removed Modern UI page and added Caxton page